+++ core/domain/src/main/kotlin/org/michaelbel/movies/domain/usecase/MovieDetailsCase.kt
@@ -8,7 +8,7 @@ import org.michaelbel.movies.entities.lce.ScreenState
 class MovieDetailsCase @Inject constructor(
     private val movieInteractor: MovieInteractor
 ) {
-    suspend operator fun invoke(movieId: Long): ScreenState {
+    suspend operator fun invoke(movieId: Int): ScreenState {
         movieInteractor.movieDetails(movieId).handle(
             success = { movieDetailsData ->
                 return ScreenState.Content(movieDetailsData)
